Refactor command_processor to be defined at driver creation time. Allows separate command processors for different specs.
