+++ lib/sync_core.dart
@@ -17,19 +17,20 @@ library webdriver.sync_core;
 import 'dart:collection' show UnmodifiableMapView;
 
 import 'package:webdriver/src/sync/capabilities.dart' show Capabilities;
-import 'package:webdriver/src/sync/command_processor.dart'
-    show CommandProcessor;
 import 'package:webdriver/src/sync/web_driver.dart' show WebDriver;
+
+import 'package:webdriver/src/sync/json_wire_spec/command_processor.dart'
+  as jwireCommand;
 import 'package:webdriver/src/sync/json_wire_spec/web_driver.dart' as jwire;
 
 export 'package:webdriver/src/sync/alert.dart';
 export 'package:webdriver/src/sync/capabilities.dart';
 export 'package:webdriver/src/sync/command_event.dart';
-export 'package:webdriver/src/sync/command_processor.dart';
 export 'package:webdriver/src/sync/common.dart';
 export 'package:webdriver/src/sync/common_spec/cookies.dart';
 export 'package:webdriver/src/sync/common_spec/navigation.dart';
 export 'package:webdriver/src/sync/exception.dart';
+export 'package:webdriver/src/sync/json_wire_spec/command_processor.dart';
 export 'package:webdriver/src/sync/json_wire_spec/keyboard.dart';
 export 'package:webdriver/src/sync/json_wire_spec/logs.dart';
 export 'package:webdriver/src/sync/json_wire_spec/mouse.dart';
@@ -43,8 +44,7 @@ final Uri defaultUri = Uri.parse('http://127.0.0.1:4444/wd/hub/');
 
 //TODO(staats): when W3C spec created, infer spec during WebDriver creation.
 
-WebDriver createDriver(CommandProcessor processor,
-    {Uri uri, Map<String, dynamic> desired}) {
+WebDriver createDriver({Uri uri, Map<String, dynamic> desired}) {
   if (uri == null) {
     uri = defaultUri;
   }
@@ -53,6 +53,7 @@ WebDriver createDriver(CommandProcessor processor,
     desired = Capabilities.empty;
   }
 
+  final processor = new jwireCommand.JsonWireCommandProcessor();
   Map response = processor.post(
       uri.resolve('session'), {'desiredCapabilities': desired},
       value: false) as Map<String, dynamic>;
@@ -60,14 +61,15 @@ WebDriver createDriver(CommandProcessor processor,
       new UnmodifiableMapView(response['value'] as Map<String, dynamic>));
 }
 
-WebDriver fromExistingSession(CommandProcessor processor, String sessionId,
+WebDriver fromExistingSession(String sessionId,
     {Uri uri}) {
   if (uri == null) {
     uri = defaultUri;
   }
 
+  final processor = new jwireCommand.JsonWireCommandProcessor();
   var response =
       processor.get(uri.resolve('session/$sessionId')) as Map<String, dynamic>;
-  return new jwire.JsonWireWebDriver(
-      processor, uri, sessionId, new UnmodifiableMapView(response));
+  return new jwire.JsonWireWebDriver(processor,
+      uri, sessionId, new UnmodifiableMapView(response));
 }
